requirements.txt
fastapi==0.124.0
pydantic==2.12.5
python-dotenv==1.2.1
uvicorn[standard]